crawler/scrapy_crawler/requirements.txt
# Scrapy爬虫依赖

# Scrapy核心
scrapy==2.11.0
scrapy-redis==0.7.3

# 数据处理
pandas==2.0.3
numpy==1.24.3

# Kafka客户端
kafka-python==2.0.2

# HDFS客户端
hdfs3==0.3.1

# 数据验证
pydantic==2.3.0

# 日志
loguru==0.7.2

# 配置管理
pyyaml==6.0.1

# HTTP请求
requests==2.31.0
aiohttp==3.8.5

# HTML解析
lxml==4.9.3
beautifulsoup4==4.12.2

# 代理
requests-socks==1.0.0

# 定时任务
schedule==1.2.0

# 工具
python-dateutil==2.8.2
